requirements.txt
psycopg2~=2.9.6
Django~=4.1.9
djangorestframework~=3.14.0